requirement.txt
grpcio==1.12.0
grpcio-tools==1.12.0
protobuf==3.5.2.post1
six==1.11.0
robotframework==3.1